Adjust gpt3 chat completion svc to receive Conversation struct

+++ lib/data-filter/main.go
@@ -13,7 +13,7 @@ type Conversation struct {
 	ID          string `json:"id"`
 	Participant string `json:"participant"`
 	Message     string `json:"message"`
-	Timestamp   int64  `json:"timestamp"`
+	Timestamp   int32  `json:"timestamp"`
 }
 
 func HandleRequest(request events.APIGatewayProxyRequest) ([]Conversation, error) {
